City menu now opens when you click on a city
handleInput methods are important. They check if something has been
clicked. If so, it makes all the other handleInput methods to get
ignored. This is done since some inputs take precedence. For example, if
end turn button is clicked on, the map should not check for input.
Otherwise, things might get selected through the button.

Also, city now is a clickable rectangle, so it can be clicked on.