Resizing V2
Ok, so I used views like your supposed to for resizing.
Works fairly well.
The menu UI remains pretty fixed. The map on the other hand remains the
same size, but if you make it bigger, more of the map is visible instead
of black bars.
This behavior seems fine to me.